requirements.txt
# Copyright 2024 GridGuard-Pi5 Contributors
# Licensed under the Apache License, Version 2.0

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0

# Hardware Interfaces
adafruit-ads1x15>=2.2.0
RPi.GPIO>=0.7.1
smbus2>=0.4.0

# Web & API
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0

# Database
SQLAlchemy>=2.0.0

# Data Processing
scipy>=1.11.0
scikit-learn>=1.3.0

# Communication
PyYAML>=6.0
requests>=2.31.0
paho-mqtt>=1.6.0

# Alerts
twilio>=8.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
